// frontend/src/api/models.ts
export const enum Approaches {
    RetrieveThenRead = "rtr",
    ReadRetrieveRead = "rrr",
    ReadDecomposeAsk = "rda"
}


export type AskRequestOverrides = {
    semanticRanker?: boolean;
    semanticCaptions?: boolean;
    excludeCategory?: string;
    top?: number;
    temperature?: number;
    promptTemplate?: string;
    promptTemplatePrefix?: string;
    promptTemplateSuffix?: string;
    suggestFollowupQuestions?: boolean;
    vectorRanker?: boolean; //:vik
};


export type AskRequest = {
    question: string;
    approach: Approaches;
    overrides?: AskRequestOverrides;
};


export type AskResponse = {
    answer: string;
    thoughts: string | null;
    data_points: string[];
    Chat_Limit: number;
    Current_Chat_Length: number;
    error?: string;
};


export type ChatTurn = {
    user: string;
    bot?: string;
};


export type ChatRequest = {
    history: ChatTurn[];
    approach: Approaches;
    overrides?: AskRequestOverrides;
};